+++ attribution/src/main/java/com/affise/attribution/AffiseApi.kt
@@ -21,6 +21,7 @@ import com.affise.attribution.usecase.FirstAppOpenUseCase
 import com.affise.attribution.usecase.ImmediateSendToServerUseCase
 import com.affise.attribution.usecase.PreferencesUseCaseImpl
 import com.affise.attribution.usecase.PushTokenUseCase
+import com.affise.attribution.usecase.SendDataToServerUseCase
 import com.affise.attribution.usecase.SendGDPREventUseCaseImpl
 import com.affise.attribution.usecase.StoreInstallReferrerUseCase
 import com.affise.attribution.webBridge.WebBridgeManager
@@ -53,8 +54,7 @@ internal interface AffiseApi {
     val postBackModelFactory: PostBackModelFactory
     val debugValidateUseCase: DebugValidateUseCase
     val debugNetworkUseCase: DebugNetworkUseCase
+    val sendDataToServerUseCase: SendDataToServerUseCase
     val immediateSendToServerUseCase: ImmediateSendToServerUseCase
     val pushTokenUseCase: PushTokenUseCase
-
-    fun isInitialized(): Boolean
 }